style: rename 'thf_climate' directory to 'src'.
