+++ tests/test_gee_basic.py
@@ -1,4 +1,4 @@
-from thf_climate.gee.auth import authenticate
+from src.gee.auth import authenticate
 
 
 def test_gee_authenticate_basic():
